# Fix 0.9.18 caching/performance regressions