--- a/src/main/scala/scalafix/internal/sbt/BlockingCache.scala
+++ b/src/main/scala/scalafix/internal/sbt/BlockingCache.scala
@@ -1,21 +1,22 @@
 package scalafix.internal.sbt
 
-import scala.collection.mutable
+import java.{util => ju}
 
 /** A basic thread-safe cache without any eviction. */
 class BlockingCache[K, V] {
-  private val underlying = new mutable.HashMap[K, V]
+
+  // Number of keys is expected to be very small so the global lock should not be a bottleneck
+  private val underlying = ju.Collections.synchronizedMap(new ju.HashMap[K, V])
 
   /**
     * @param value By-name parameter evaluated when the key if missing. Value computation is guaranteed
     *              to be called only once per key across all invocations.
     */
-  def getOrElseUpdate(key: K, value: => V): V = {
-    // ConcurrentHashMap does not guarantee that there is only one evaluation of the value, so
-    // we use our own (global) locking, which is OK as the number of keys is expected to be
-    // very small (bound by the number of projects in the sbt build).
-    underlying.synchronized {
-      underlying.getOrElseUpdate(key, value)
-    }
-  }
+  def getOrElseUpdate(key: K, value: => V): V =
+    underlying.computeIfAbsent(
+      key,
+      new ju.function.Function[K, V] {
+        override def apply(k: K): V = value
+      }
+    )
 }