* The first commit fixes an embarrassingly massive performance regression introduced by 2ca68e4: the comment/assumption in https://github.com/scalacenter/sbt-scalafix/pull/126#discussion_r439692348 was wrong, causing artifact resolution & classloader creation to happen for each invocation/project/configuration, which caused slow startup and slower rule execution (JIT not warmed up). It might also fix https://github.com/scalacenter/scalafix/issues/1196#issuecomment-654744979.
* The second commit brings back the 0.9.17 behavior of looking up last commits for the parser at the first `scalafix` invocation (changed also in 2ca68e4), to reduce the risk of JGit stuff blocking the initialization like https://github.com/scalacenter/scalafix/issues/1196. Related to https://github.com/scalacenter/sbt-scalafix/pull/141.
* The two last commits fix a minor known regression introduced by https://github.com/scalacenter/sbt-scalafix/commit/d81dd7266231e70af475551b563b21ec75c89482 (see message of that commit).

We'll probably need some performance tests to catch those regressions in the future...

### Testing on a real project

I have timed the initialization overhead of `scalafix` by running a warm `all scalafix test:scalafix it:scalafix` (third invocation with `scalafixCaching := true`) on a build with 25 sub-projects which have Scalafix enabled.

* With `scalafixDependencies`
  * 0.9.16: ~10s
  * 0.9.18: ~50s
  * This PR: ~8s

* With `scalafixDependencies` & `dependsOn(customRules % ScalafixConfig)`
  * 0.9.16: ~10s
  * 0.9.18: ~50s
  * This PR: ~9s

The impact of (not) JITing the rules is not exposed by this test, but rule execution should be back to pre-0.9.18 level as well.